docs: moving and slimming down comments
